video: pass mmap's off_t offset as isize, not i64
The v4l2 and framebuffer raw-FFI blocks declared mmap's final argument as
`i64`. `off_t` is `long` — pointer-width — so on 32-bit ARM the 64-bit
declaration mis-passes the offset: AAPCS aligns the i64 to a wider stack
slot than a 32-bit off_t, so glibc reads a misaligned offset and fails mmap
in userspace before the syscall ever runs. The V4L2 capture buffers then
never map and streaming can't start (`QUERYBUF` succeeds, mmap does not).
Use `isize`, which is `long` on every arch — unchanged on 64-bit, correct
on 32-bit. Mirrors the existing pointer-width note on the `ioctl` request.

+++ crates/yscv-video/src/framebuffer.rs
@@ -23,7 +23,16 @@ mod linux_impl {
         // Declaring it `u64` here disagreed with the same symbol in `v4l2.rs`
         // and passed a wrong-width argument on a 32-bit board.
         fn ioctl(fd: i32, request: usize, ...) -> i32;
-        fn mmap(addr: *mut u8, len: usize, prot: i32, flags: i32, fd: i32, offset: i64) -> *mut u8;
+        // `off_t` is `long`, same story as `request` above — pointer-width, so
+        // `i64` mis-passes it on 32-bit ARM. Use `isize`.
+        fn mmap(
+            addr: *mut u8,
+            len: usize,
+            prot: i32,
+            flags: i32,
+            fd: i32,
+            offset: isize,
+        ) -> *mut u8;
         fn munmap(addr: *mut u8, len: usize) -> i32;
     }
